Begin user formatting
