+++ src/ms.c
@@ -1,4 +1,3 @@
-#include "utils.h"
 #include "args.h"
 
 int main(int argc, char** argv) {
