+++ src/tuners/batch.c
@@ -1,4 +1,3 @@
-#include "utils.h"
 #include "thread.h"
 #include "args.h"
